requirements.txt
alembic==1.13.0
bcrypt==4.1.1
email-validator==2.1.0.post1
fastapi==0.104.1
PyJWT==2.8.0
python-dotenv==1.0.0
uvicorn==0.24.0.post1
